+++ R/utils.R
@@ -32,7 +32,7 @@
 
 ### Set to NA if standardised values of 'estvarname' or 'se' > 'max' (grouped by 'methodvar', 'by') using either normal ((x - mean) / stdev) or robust ((x - median) / IQR) standardisation
 #' @keywords internal
-.dropbig <- function(data, estvarname, se, methodvar, by, max, semax, robust, internal = TRUE) {
+.dropbig <- function(data, estvarname, se = NULL, methodvar, by, max, semax, robust, internal = TRUE) {
   splt_c <- c(methodvar, by)
   if (length(splt_c) > 0) {
     idata <- .split_by(data, splt_c)
@@ -43,33 +43,43 @@
     tmp <- idata[[i]]
     if (robust) {
       tmp[[paste0(".", estvarname, ".std")]] <- (tmp[[estvarname]] - median(tmp[[estvarname]], na.rm = TRUE)) / (fivenum(tmp[[estvarname]], na.rm = TRUE)[4] - fivenum(tmp[[estvarname]], na.rm = TRUE)[2])
-      tmp[[paste0(".", se, ".std")]] <- (tmp[[se]] - median(tmp[[se]], na.rm = TRUE)) / (fivenum(tmp[[se]], na.rm = TRUE)[4] - fivenum(tmp[[se]], na.rm = TRUE)[2])
+      if (!is.null(se)) tmp[[paste0(".", se, ".std")]] <- (tmp[[se]] - median(tmp[[se]], na.rm = TRUE)) / (fivenum(tmp[[se]], na.rm = TRUE)[4] - fivenum(tmp[[se]], na.rm = TRUE)[2])
     } else {
       tmp[[paste0(".", estvarname, ".std")]] <- (tmp[[estvarname]] - mean(tmp[[estvarname]], na.rm = TRUE)) / sqrt(var(tmp[[estvarname]], na.rm = TRUE))
-      tmp[[paste0(".", se, ".std")]] <- (tmp[[se]] - mean(tmp[[se]], na.rm = TRUE)) / sqrt(var(tmp[[se]], na.rm = TRUE))
+      if (!is.null(se)) tmp[[paste0(".", se, ".std")]] <- (tmp[[se]] - mean(tmp[[se]], na.rm = TRUE)) / sqrt(var(tmp[[se]], na.rm = TRUE))
     }
     tmp
   })
   odata <- .br(odata)
   if (internal) {
     odata[[estvarname]] <- ifelse(abs(odata[[paste0(".", estvarname, ".std")]]) > max, NA, odata[[estvarname]])
-    odata[[se]] <- ifelse(abs(odata[[paste0(".", se, ".std")]]) > semax, NA, odata[[se]])
     odata[[paste0(".", estvarname, ".std")]] <- NULL
-    odata[[paste0(".", se, ".std")]] <- NULL
+    if (!is.null(se)) {
+      odata[[se]] <- ifelse(abs(odata[[paste0(".", se, ".std")]]) > semax, NA, odata[[se]])
+      odata[[paste0(".", se, ".std")]] <- NULL
+    }
   } else {
-    odata[[".dropbig"]] <- ifelse(abs(odata[[paste0(".", estvarname, ".std")]]) > max | abs(odata[[paste0(".", se, ".std")]]) > semax, TRUE, FALSE)
+    if (!is.null(se)) {
+      odata[[".dropbig"]] <- ifelse(abs(odata[[paste0(".", estvarname, ".std")]]) > max | abs(odata[[paste0(".", se, ".std")]]) > semax, TRUE, FALSE)
+      odata[[paste0(".", se, ".std")]] <- NULL
+    } else {
+      odata[[".dropbig"]] <- ifelse(abs(odata[[paste0(".", estvarname, ".std")]]) > max, TRUE, FALSE)
+    }
     odata[[paste0(".", estvarname, ".std")]] <- NULL
-    odata[[paste0(".", se, ".std")]] <- NULL
   }
   return(odata)
 }
 
 ### Set both est, se to NA if any of the two is NA
 #' @keywords internal
-.na_pair <- function(data, estvarname, se) {
-  toNA <- (is.na(data[[estvarname]]) | is.na(data[[se]]))
+.na_pair <- function(data, estvarname, se = NULL) {
+  if (!is.null(se)) {
+    toNA <- (is.na(data[[estvarname]]) | is.na(data[[se]]))
+    data[[se]][toNA] <- NA
+  } else {
+    toNA <- (is.na(data[[estvarname]]))
+  }
   data[[estvarname]][toNA] <- NA
-  data[[se]][toNA] <- NA
   data
 }
 
@@ -78,7 +88,7 @@
 .describe <- function(x, ref, level) {
   description_df <- data.frame(
     stat = c("nsim", "thetamean", "thetamedian", "se2mean", "se2median", "bias", "empse", "relprec", "mse", "modelse", "relerror", "cover", "becover", "power"),
-    description = c("Non-missing point estimates/standard errors", "Average point estimate", "Median point estimate", "Average standard error", "Median standard error", "Bias in point estimate", "Empirical standard error", paste("% gain in precision relative to method", ref), "Mean squared error", "Model-based standard error", "Relative % error in standard error", paste("Coverage of nominal", sprintf("%.0f%%", 100 * (level)), "confidence interval"), paste("Bias-eliminated coverage of nominal", sprintf("%.0f%%", 100 * (level)), "confidence interval"), paste("Power of", sprintf("%.0f%%", 100 * (1 - level)), "level test")),
+    description = c("Non-missing point estimates/standard errors", "Average point estimate", "Median point estimate", "Average variance", "Median variance", "Bias in point estimate", "Empirical standard error", paste("% gain in precision relative to method", ref), "Mean squared error", "Model-based standard error", "Relative % error in standard error", paste("Coverage of nominal", sprintf("%.0f%%", 100 * (level)), "confidence interval"), paste("Bias-eliminated coverage of nominal", sprintf("%.0f%%", 100 * (level)), "confidence interval"), paste("Power of", sprintf("%.0f%%", 100 * (1 - level)), "level test")),
     stringsAsFactors = FALSE
   )
   x <- merge(x, description_df, by = "stat")