+++ tests/testthat/test-multisimsum.R
@@ -7,6 +7,7 @@ testthat::test_that("multisimsum prints ok", {
   testthat::expect_output(print(multisimsum(data = frailty, par = "par", true = c(trt = -0.50, fv = 0.75), estvarname = "b", se = "se", methodvar = "model")))
   testthat::expect_output(print(multisimsum(data = frailty, par = "par", true = c(trt = -0.50, fv = 0.75), estvarname = "b", se = "se")))
   testthat::expect_output(print(multisimsum(data = frailty, par = "par", true = c(trt = -0.50, fv = 0.75), estvarname = "b", se = "se", methodvar = "model", by = "fv_dist", control = list(mcse = FALSE))))
+  testthat::expect_output(print(multisimsum(data = frailty, par = "par", estvarname = "b", se = "se")))
 })
 
 testthat::test_that("multisimsum returns an object of class multisimsum", {
@@ -22,38 +23,33 @@ testthat::test_that("summ slot of a multisimsum object is a data.frame", {
 })
 
 testthat::test_that("not passing estvarname throws an error", {
-  testthat::expect_error({
-    data("frailty")
-    ms <- multisimsum(data = frailty, par = "par", true = c(trt = -0.50, fv = 0.75), se = "se", methodvar = "model", by = "fv_dist")
-  }, 'argument "estvarname" is missing, with no default')
-})
-
-testthat::test_that("not passing true throws an error", {
-  testthat::expect_error({
-    data("frailty")
-    ms <- multisimsum(data = frailty, par = "par", estvarname = "b", se = "se", methodvar = "model", by = "fv_dist")
-  }, 'argument "true" is missing, with no default')
-})
-
-testthat::test_that("not passing se throws an error", {
-  testthat::expect_error({
-    data("frailty")
-    ms <- multisimsum(data = frailty, par = "par", true = c(trt = -0.50, fv = 0.75), estvarname = "b", methodvar = "model", by = "fv_dist")
-  }, 'argument "se" is missing, with no default')
+  testthat::expect_error(
+    {
+      data("frailty")
+      ms <- multisimsum(data = frailty, par = "par", true = c(trt = -0.50, fv = 0.75), se = "se", methodvar = "model", by = "fv_dist")
+    },
+    'argument "estvarname" is missing, with no default'
+  )
 })
 
 testthat::test_that("specifying ref and not methodvar throws a warning", {
-  testthat::expect_warning({
-    data("frailty")
-    ms <- multisimsum(data = frailty, par = "par", true = c(trt = -0.50, fv = 0.75), estvarname = "b", se = "se", ref = "Cox, Gamma", by = "fv_dist")
-  }, "'ref' method is specified while 'methodvar' is not: 'ref' will be ignored")
+  testthat::expect_warning(
+    {
+      data("frailty")
+      ms <- multisimsum(data = frailty, par = "par", true = c(trt = -0.50, fv = 0.75), estvarname = "b", se = "se", ref = "Cox, Gamma", by = "fv_dist")
+    },
+    "'ref' method is specified while 'methodvar' is not: 'ref' will be ignored"
+  )
 })
 
 testthat::test_that("specifying methodvar and not ref shows a message", {
-  testthat::expect_message({
-    data("frailty")
-    ms <- multisimsum(data = frailty, par = "par", true = c(trt = -0.50, fv = 0.75), estvarname = "b", se = "se", methodvar = "model", by = "fv_dist")
-  }, "'ref' method was not specified, Cox, Gamma set as the reference")
+  testthat::expect_message(
+    {
+      data("frailty")
+      ms <- multisimsum(data = frailty, par = "par", true = c(trt = -0.50, fv = 0.75), estvarname = "b", se = "se", methodvar = "model", by = "fv_dist")
+    },
+    "'ref' method was not specified, Cox, Gamma set as the reference"
+  )
 })
 
 testthat::test_that("running multisimsum on frailty return summaries of the correct dimension", {
@@ -118,4 +114,22 @@ testthat::test_that("multisimsum argument checks", {
   testthat::expect_error(object = rsimsum::multisimsum(data = frailty, par = NULL, true = c(trt = -0.50, fv = 0.75), estvarname = "b", se = "se", methodvar = "model", by = "fv_dist"))
   testthat::expect_error(object = rsimsum::multisimsum(data = frailty, par = "par", true = 1, estvarname = "b", se = "se", methodvar = "model", by = "fv_dist"))
   testthat::expect_error(object = rsimsum::multisimsum(data = frailty, par = "par", true = TRUE, estvarname = "b", se = "se", methodvar = "model", by = "fv_dist"))
+})
+
+testthat::test_that("multisimsum without 'true' does not compute bias, cover, mse", {
+  data("frailty")
+  s <- multisimsum(data = frailty, par = "par", estvarname = "b", se = "se", methodvar = "model", by = "fv_dist")
+  testthat::expect_false(object = any(c("bias", "cover", "mse") %in% s$summ$stat))
+})
+
+testthat::test_that("simsum without 'se' does not compute se2mean, se2median, modelse, relerror, cover, becover, power", {
+  data("frailty")
+  s <- multisimsum(data = frailty, par = "par", estvarname = "b", true = c(trt = -0.50, fv = 0.75), methodvar = "model", by = "fv_dist")
+  testthat::expect_false(object = any(c("se2mean", "se2median", "modelse", "relerror", "cover", "becover", "power") %in% s$summ$stat))
+})
+
+testthat::test_that("simsum without 'se' nor 'true' does not compute se2mean, se2median, modelse, relerror, cover, becover, power, bias, mse", {
+  data("frailty")
+  s <- multisimsum(data = frailty, par = "par", estvarname = "b", methodvar = "model", by = "fv_dist")
+  testthat::expect_false(object = any(c("se2mean", "se2median", "modelse", "relerror", "cover", "becover", "power", "bias", "mse") %in% s$summ$stat))
 })